Feature Flag with SwiftUI
Feature: A Feature class is implemented with two functionality:

A feature flag is implemented to return a nick name for username that has been entered by user.
A feature flag capitalizeNickname is implememnted which returns if the nick name has to be capitalized or not.
Both of the above feature flags are used in combination to get nickname and capitalize it as per the feature flag state.